colorscheme: Massive refactor to account for bold
I want better control over colors on low-color terminals like the linux
console.  To do that I need to know when to enable bold to be able to
access (especially) bright white and bright black.  This move more data
into the Color class and tries to enforce a more functional style.