status-lib: Use special FIFO mode to allow writing w/o reader
Linux allows FIFOs to be opened in read-write mode in which case my
monitor processes can continue to write to a fifo after the reader has
closed its FD, and messages are not lost.  This is a hacky way to get
simple message passing between my short-lived processes.

See: fifo(7)