vim: relativenumber is hopelessly broken
While [patch
8.1.0374](https://github.com/vim/vim/commit/bd9a53c06c8869ad811cb3dd01a309c9be7d7a63)
seems to fix scrolling, relativenumber is still painfully slow when
typing code under the default regexpengine.  Setting regexpengine=1
helps, but doesn't solve the problem, it only hides it with a faster
regex implementation.  The only use of the relativenumber line was for
jumps and the occasional line changes/deletions.  I should use / and ?
searches instead.