Make awk code posix compatible
The character class `\s` is not available in POSIX awk and compatibles
like mawk, installed on Ubuntu by default according to #1.  We could use
`[[:space:]]` instead and that works, but it is probably more correct
not to use a regular expression at all when our intent is to match
exactly on the first field.

Closes #1