sway-layout-manager: Fix 'balance' on multi-output
The balance function failed when rebuilding the master stack, resulting
in a window being moved from the right to the left output.  Issue a
command to move the window back to the original output, which results in
the master rebuild I wanted anyway.  Seems to work reliably on both left
and right outputs.  Don't have a way to test a middle output.