sway-layout-manager: Fixes for Sway 1.6
Sway 1.6 optimized away splits from containers with only one window.
The result is we need to use more `layout splitX` instead of `splitX` to
assert what the layout should be rather than what it is.  Then, when a
second window opens, it takes on the specified layout.  At least, that's
how I think this is working now--the documentation and behavior aren't
entirely clear.

The biggest change that was needed was to the `flip` function.  Under
Sway 1.6, there does not seem to be a way to change a top-level
container's layout when it has multiple child containers--it just
changes the layout of the first child container.  I tried many
combinations of commands.  Change the filp algorithm to move the second
container into the first before toggling the outer container layout.
Then move the second container out of the first.  When executed as one
swaymsg, all these moves seem to be atomic.