picom: Run on VMware and restart when Chrome kills it
Compositing with GLX kicks the VMware display into 60 fps mode. Without
it, X will render 60 fps, but VMware will only display 30 of them. Picom
causes the Chrome window borders to be rendered transparent due to its
use of ARGB visuals; however, if picom is not running when the first
Chromium instance launched, it seems not to enable the ARGB visuals and
picom can be rerun as normal. The "Restart=always" policy acomplishes
this with a corresponding pkill in the wrapper scripts.