sway: Move idle inhibitor to scratchpad
When sway starts without connected display, the idler workspace that the
idle inhibitor used to be moved to would be displayed upon connecting
the display. The scratchpad is never shown so moving the inhibitor
window to it ensures it is always hidden. The scratchpad is not
otherwise important to my workflow.

Closes #6