git: Disable directory ownership check
This recent git behavior change breaks Puppet's vcsrepo module < 5.0.0.
Versions >= 5.0.0 are hopelessly broken on Windows. All this is relating
to switching to another user to run git, something that now needs to be
performed as long as this safety mechanism is in place, and something
that can't be done easily on Windows. Just disable the check.