zsh: Reload VS code shell integration
VS Code only injects the shell integration environment variables on the
first invocation of the terminal. If we reload the shell, re-exec it,
source the zshrc, etc., we lose the shell integration, so load it
explicitly. This seems to have to come after the prompt initialization
to work, which kinda makes sense. Just throw it in at the end as a
special thing.